remove modify category dialog and associated stuff
