--- Src/MoneyFox.Windows/Views/Dialogs/ModifyCategoryDialog.xaml
@@ -1,26 +1,0 @@
-﻿<ContentDialog x:Class="MoneyFox.Windows.Views.Dialogs.ModifyCategoryDialog"
-               xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-               xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-               xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-               xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-               xmlns:core="using:Microsoft.Xaml.Interactions.Core"
-               xmlns:interactivity="using:Microsoft.Xaml.Interactivity"
-               mc:Ignorable="d"
-               Title="{Binding Title}"
-               x:Uid="ModifyCategoryDialog"
-               PrimaryButtonText="done"
-               SecondaryButtonText="cancel"
-               PrimaryButtonCommand="{Binding DoneCommand}">
-
-    <interactivity:Interaction.Behaviors>
-        <core:EventTriggerBehavior EventName="Loaded">
-            <core:InvokeCommandAction Command="{Binding LoadedCommand, Mode=OneWay}" />
-        </core:EventTriggerBehavior>
-    </interactivity:Interaction.Behaviors>
-
-    <StackPanel VerticalAlignment="Stretch" HorizontalAlignment="Stretch">
-        <TextBox Header="name" x:Uid="Name"
-                 Text="{Binding Selected.Name, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"
-                 KeyDown="TextBox_OnKeyDown" />
-    </StackPanel>
-</ContentDialog>
