+++ Src/MoneyFox.Windows/Views/SettingsView.xaml.cs
@@ -1,19 +1,10 @@
-﻿using System;
-using Windows.UI.Xaml;
-using MoneyFox.Windows.Views.Dialogs;
-
-namespace MoneyFox.Windows.Views
+﻿namespace MoneyFox.Windows.Views
 {
     public sealed partial class SettingsView
     {
         public SettingsView()
         {
             InitializeComponent();
-        }
-
-        private async void AddCategory(object sender, RoutedEventArgs e)
-        {
-            await new ModifyCategoryDialog().ShowAsync();
         }
     }
 }
